Tweak build matrix to support PHP 7.2
We need to consider 7.2 as a viable build target but for the purposes of deprecation (which is all we test for right now), we only need to care about 7.0 as our minimum and 7.2 as our (effective) maximum. We still only produce two builds but that's enough to cover all we test right now.