Commit the first part of the parser overhaul, splitting it out.
The plan is to have subclassed parsers, e.g. for posts vs signatures
vs character sheets where different tags can be used but instead of
declaring these at parse time (and thus subject to non-caching), they
should be declarable at instance time and bound so that you can't use
certain bbcode in, say, signatures that you can in posts. This commit
is just starting to explore how this is done and is mostly being done
so it's there regardless.